// src/api/generated/schemas/characterListItemBodyType.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * ClothingFit API
 * ClothingFit 백엔드 API 문서
 * OpenAPI spec version: 1.0.0
 */

export type CharacterListItemBodyType = typeof CharacterListItemBodyType[keyof typeof CharacterListItemBodyType];


export const CharacterListItemBodyType = {
  SLIM: 'SLIM',
  NORMAL: 'NORMAL',
  OVERWEIGHT: 'OVERWEIGHT',
  OBESE: 'OBESE',
} as const;