// src/api/generated/schemas/characterListItemGender.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * ClothingFit API
 * ClothingFit 백엔드 API 문서
 * OpenAPI spec version: 1.0.0
 */

export type CharacterListItemGender = typeof CharacterListItemGender[keyof typeof CharacterListItemGender];


export const CharacterListItemGender = {
  MALE: 'MALE',
  FEMALE: 'FEMALE',
} as const;